test(profiling): move always-excluded-modules tests to test_profiling_config
- Move test_always_excluded_modules_cannot_be_overridden from test_threading.py
- Add test_always_excluded_modules_contains_required_entries (direct membership assertion)
- Expose _ALWAYS_EXCLUDED_MODULES in _lock.pyi stub so mypy can see it
